pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "desk-grade-ready"
version = "0.1.0"
description = "Desk-Grade Trading System - Sistema profesional de trading automatizado"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Desk-Grade Team"}
]
keywords = ["trading", "risk-management", "automated-trading"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "psycopg[binary]==3.2.3",
    "numpy==1.26.4",
    "pandas==2.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
desk-grade-risk-cycle = "scripts.run_risk_cycle:main"
desk-grade-scheduler = "scripts.scheduler:main"
desk-grade-health = "scripts.health_check:main"
desk-grade-status = "scripts.status:main"
desk-grade-seed = "scripts.seed_data:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
